encapsulate single_winner implementation
